update fixation grid comparison functions
moved to .stats.py
