+++ PyEyeSim/__init__.py
@@ -20,7 +20,7 @@ from matplotlib.patches import Ellipse
 import platform
 #%%
 from .visualhelper import VisBinnedProg,PlotDurProg,JointBinnedPlot,MeanPlot,draw_ellipse,HistPlot
-from .scanpathsimhelper import AOIbounds,CreatAoiRects,Rect,SaccadeLine,CalcSim ,CheckCorr
+from .scanpathsimhelper import AOIbounds,CreatAoiRects,Rect,SaccadeLine,CalcSim, CheckCoor
 from .statshelper import SaliencyMapFilt,SaccadesTrial,ScanpathL,StatEntropy
 
 
@@ -28,7 +28,7 @@ class EyeData:
 	
     from ._visuals import VisScanPath,MySaccadeVis,VisLOOHMM,VisHMM,MyTrainTestVis
     from ._dataproc import GetParams,GetStimuli,GetFixationData,GetDurations,GetGroups,GetCats,GetSaccades,SaccadeSel,GetEntropies,InferSize,Heatmap
-    from ._stats import AngleCalc,AngtoPix,PixdoDeg,Entropy,FixDurProg,BinnedCount
+    from ._stats import AngleCalc,AngtoPix,PixdoDeg,Entropy,FixDurProg,BinnedCount,GetInddiff,GetInddiff_v2,RunDiffDivs,GetBinnedStimFixS,StatPDiffInd2,StatPDiffInd1,CalcStatPs
     try: 
     	from ._hmm import DataArrayHmm,MyTrainTest,FitLOOHMM,FitVisHMM,FitVisHMMGroups,HMMSimPipeline
     except:
@@ -409,8 +409,8 @@ class EyeData:
         WhichAOIH=np.zeros(NFix)
         WhichAOIV=np.zeros(NFix)
         for x in range(NFix):
-            WhichAOIH[x]=CheckCor(AOIboundsH,FixTrialX[x]) # store which horizontal AOI each fixation is
-            WhichAOIV[x]=CheckCor(AOIboundsV,FixTrialY[x]) # store which vertical AOI each fixation is
+            WhichAOIH[x]=CheckCoor(AOIboundsH,FixTrialX[x]) # store which horizontal AOI each fixation is
+            WhichAOIV[x]=CheckCoor(AOIboundsV,FixTrialY[x]) # store which vertical AOI each fixation is
     
         WhichAOI=np.zeros(NFix)
         WhichAOI[:]=np.NAN
@@ -424,98 +424,6 @@ class EyeData:
     
     
    
-    
-    def StatPDiffInd1(self,statPMat):
-        StatIndDiff=np.zeros(((self.np,self.ns,self.ns)))
-        for cp,p in enumerate(self.stimuli):   
-            for cs1,s1 in enumerate(self.subjects):
-                for cs2,s2 in enumerate(self.subjects):
-                      StatIndDiff[cp,cs1,cs2]=np.nansum((statPMat[cs1,cp,:,:]-statPMat[cs2,cp,:,:])**2)
-        return StatIndDiff
-     
-    def StatPDiffInd2(self,BindAll):
-        StatIndDiff=np.zeros(((self.np,self.ns,self.ns)))
-        for cp,p in enumerate(self.stimuli):   
-            for cs1,s1 in enumerate(self.subjects):
-                for cs2,s2 in enumerate(self.subjects):
-                     StatIndDiff[cp,cs1,cs2]=np.nansum((BindAll[cp][cs1,:,:]-BindAll[cp][cs2,:,:])**2)
-        return StatIndDiff
-    
-    
-    def GetInddiff(self,nHor,nVer,Vis=0,zscore=0,InferS=1):
-        ''' N DIVISION BASED. calculate individual similarity between all pairs of participants for all stimuli, for a given division'''
-        statPMat,statEntropyMat=self.CalcStatPs(nHor,nVer,InferS=InferS)
-     
-        Inddiff=self.StatPDiffInd1(statPMat)
-        Indmean=np.nanmean(Inddiff,2)
-        SD=np.nanstd(Indmean,1)
-        Indmean=np.nanmean(Indmean,1)
-        if Vis:
-            #plt.errorbar(np.arange(self.np),Indmean,SD,marker='o',linestyle='none')
-            if zscore:
-                plt.scatter(np.arange(self.np),(Indmean-np.mean(Indmean))/np.std(Indmean),marker='o')
-            else:
-                plt.scatter(np.arange(self.np),Indmean,marker='o')
-            plt.xticks(np.arange(self.np),self.stimuli,rotation=80,fontsize=12)
-            plt.xlabel('Stimuli',fontsize=14)
-            if zscore==1:
-                plt.ylabel('fixation map relative difference',fontsize=14)
-            else:
-                plt.ylabel('fixation map difference',fontsize=14)
-        return Indmean
-      
-    
-    def GetInddiff_v2(self,size=50,Vis=0,fixs=0):
-        ''' PIXEl; NUMBER BASED; calculate individual similarity between all pairs of participants for all stimuli, for a given division'''
-        statPMat=self.GetBinnedStimFixS(size=size,fixs=fixs)
-        Inddiff=self.StatPDiffInd2(statPMat)
-        Indmean=np.nanmean(Inddiff,2)
-        SD=np.nanstd(Indmean,1)
-        Indmean=np.nanmean(Indmean,1)
-        if Vis:
-            #plt.errorbar(np.arange(self.np),Indmean,SD,marker='o',linestyle='none')
-            plt.scatter(np.arange(self.np),Indmean,marker='o')
-            plt.xticks(np.arange(self.np),self.stimuli,rotation=80,fontsize=12)
-            plt.xlabel('Stimuli',fontsize=14)
-            plt.ylabel('fixation map difference',fontsize=14)
-        return Indmean
-      
-
-    def GetBinnedStimFixS(self,size=50,fixs=1):
-        ''' fixs=1: use full stimulus area
-        fixs=0: use active area with 99% fixations '''
-        BindAll=[]
-        for cp,p in enumerate(self.stimuli):
-            Fixcounts=self.FixCountCalc(p,CutAct=0)
-            print('array size',np.round((Fixcounts.nbytes/1024)/1024,2),'MB')
-            binIndC=self.BinnedCount(Fixcounts[0],p,fixs=fixs,binsize_h=size)
-            BinDims=np.shape(binIndC)
-           # print(cp,BinDims)
-            BindAll.append(np.zeros(((self.ns,BinDims[0],BinDims[1]))))
-            for cs,s in enumerate(self.subjects):
-                BindAll[cp][cs,:,:]=self.BinnedCount(Fixcounts[cs],p,fixs=fixs,binsize_h=size)    
-                BindAll[cp][cs,:,:]/=np.sum(BindAll[cp][cs,:,:])
-        return BindAll
-    
-    
-    def RunDiffDivs(self,mindiv,maxdiv,Vis=1):
-        ''' run grid based fixation map comparison from 
-        mindiv*mindiv 
-        to maxdiv *maxdiv number of divisions
-        vis=1: visualized mean similarity'''
-        if Vis:
-            fig,ax=plt.subplots()
-           # plt.figure()
-        DiffsRaw=np.zeros((self.np,maxdiv-mindiv))
-        DiffsZscore=np.zeros((self.np,maxdiv-mindiv))
-        for cdiv,divs in enumerate(np.arange(mindiv,maxdiv)):
-            DiffsRaw[:,cdiv]=self.GetInddiff(divs,divs,Vis=Vis,zscore=1)
-            DiffsZscore[:,cdiv]=(DiffsRaw[:,cdiv]-np.mean(DiffsRaw[:,cdiv]))/np.std(DiffsRaw[:,cdiv])
-        if Vis:
-            ax.errorbar(np.arange(self.np),np.mean(DiffsZscore,1),np.std(DiffsZscore,1),linestyle='none',color='k',marker='o',markersize=5)
-        return DiffsZscore,DiffsRaw
-    
-
