+++ PyEyeSim/_dataproc.py
@@ -5,7 +5,7 @@ from scipy import stats,ndimage
 import pandas as pd
 import matplotlib.pyplot as plt
 from .statshelper import SaliencyMapFilt,SaccadesTrial,ScanpathL,StatEntropy
-from .scanpathsimhelper import AOIbounds,CreatAoiRects,Rect,SaccadeLine,CalcSim ,CheckCorr
+from .scanpathsimhelper import AOIbounds,CreatAoiRects,Rect,SaccadeLine,CalcSim ,CheckCoor
 import platform
